Add actual cli function inside the package
